✨ 이미지 업로드 커스텀훅 및 이미지압축 모듈 추가
