+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.5.1",
+    "browser-image-compression": "^2.0.2",
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "helmet": "^7.0.0",
     "html2canvas": "^1.4.1",
